backend/requirements.txt
# ============================================
# LAIKA CLUB - DEPENDENCIAS PYTHON
# ============================================

# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
SQLAlchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3

# QR Codes (para boletos)
qrcode[pil]==7.4.2
Pillow==10.2.0

# Excel/CSV (para exportaciones)
openpyxl==3.1.2
pandas==2.2.0

# Testing (opcional)
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0
psutil==5.9.8
